fix merge conflicts.

+++ Src/MoneyFox.Application/Accounts/Queries/GetIncludedAccountBalanceSummary/GetIncludedAccountBalanceSummaryQuery.cs
@@ -7,9 +7,9 @@ using MoneyFox.Application.QueryObjects;
 
 namespace MoneyFox.Application.Accounts.Queries.GetIncludedAccountBalanceSummary
 {
-    public class GetIncludedAccountBalanceSummaryQuery : IRequest<double>
+    public class GetIncludedAccountBalanceSummaryQuery : IRequest<decimal>
     {
-        public class Handler : IRequestHandler<GetIncludedAccountBalanceSummaryQuery, double>
+        public class Handler : IRequestHandler<GetIncludedAccountBalanceSummaryQuery, decimal>
         {
             private readonly IEfCoreContext context;
 
@@ -18,7 +18,7 @@ namespace MoneyFox.Application.Accounts.Queries.GetIncludedAccountBalanceSummary
                 this.context = context;
             }
 
-            public async Task<double> Handle(GetIncludedAccountBalanceSummaryQuery request, CancellationToken cancellationToken)
+            public async Task<decimal> Handle(GetIncludedAccountBalanceSummaryQuery request, CancellationToken cancellationToken)
             {
                 return await context.Accounts.AreNotExcluded().SumAsync(x => x.CurrentBalance, cancellationToken);
             }
