+++ Src/MoneyFox.Presentation/ViewModels/PaymentListBalanceViewModel.cs
@@ -34,7 +34,7 @@ namespace MoneyFox.Presentation.ViewModels
         ///     Calculates the sum of all accounts at the current moment.
         /// </summary>
         /// <returns>Sum of the balance of all accounts.</returns>
-        protected override async Task<double> CalculateTotalBalanceAsync()
+        protected override async Task<decimal> CalculateTotalBalanceAsync()
         {
             AccountViewModel account = mapper.Map<AccountViewModel>(await mediator.Send(new GetAccountByIdQuery(accountId)));
             return account.CurrentBalance;
@@ -45,7 +45,7 @@ namespace MoneyFox.Presentation.ViewModels
         ///     This includes all payments coming until the end of month.
         /// </summary>
         /// <returns>Balance of the selected account including all payments to come till end of month.</returns>
-        protected override async Task<double> GetEndOfMonthValueAsync()
+        protected override async Task<decimal> GetEndOfMonthValueAsync()
         {
             AccountViewModel account = mapper.Map<AccountViewModel>(await mediator.Send(new GetAccountByIdQuery(accountId)));
             return await balanceCalculationService.GetEndOfMonthBalanceForAccount(account);
